# docs/openapi.yaml
...
openapi: 3.0.0
info:
  title: Person Endpoints
  description: This is a person API from Udacity
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/Virtus5/Company/1.0.0
  description: SwaggerHub API Auto Mocking
- url: http://localhost:30001/
  description: Person API
paths:
...
  /api/persons/{personId}/connection:
    get:
      tags:
      - get connection
      summary: gets connection
      description: |
        Gets conenction
      operationId: getConnection
      parameters:
      - name: personId
        in: path
        description: ID of the user to get conenction
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Person'
        "400":
          description: bad input
        "404":
          description: not found
components:
  schemas:
    Person:
      required:
      - company_name
      - first_name
      - id
      - last_name
      type: object
      properties:
        id:
          type: integer
          example: 1
        first_name:
          type: string
          example: Thomas
        last_name:
          type: string
          example: Mueller
        company_name:
          type: string
          example: VSE AG
...